# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/users": {
      "delete": {
        "operationId": "removeUsers",
        "parameters": [
          {
            "in": "query",
            "name": "usernames",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Delete users",
        "tags": [
          "user-service"
        ]
      },
      "get": {
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PnPUser"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all users",
        "tags": [
          "user-service"
        ]
      },
      "post": {
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PnPUserCreation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Create a user",
        "tags": [
          "user-service"
        ]
      }
    },
...
    "/api/users/{username}": {
      "delete": {
        "operationId": "removeUser",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "user-service"
        ]
      },
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PnPUser"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get a user",
        "tags": [
          "user-service"
        ]
      },
      "put": {
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PnPUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Updates a user",
        "tags": [
          "user-service"
        ]
      }
    },
    "/api/users/{username}/permissions": {
      "get": {
        "operationId": "getPermissions",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/GrantedUniverseAuthorityDTO"
                      },
                      {
                        "$ref": "#/components/schemas/RoleAuthorityDTO"
                      }
                    ]
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets the permissions of a user",
        "tags": [
          "user-service"
        ]
      },
      "post": {
        "operationId": "updatePermissions",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/GrantedUniverseAuthorityDTO"
                    },
                    {
                      "$ref": "#/components/schemas/RoleAuthorityDTO"
                    }
                  ]
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Updates the permissions of a user",
        "tags": [
          "user-service"
        ]
      }
    },
    "/api/users/{username}/preferences": {
      "get": {
        "operationId": "getUserPreferences",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PnPUserPreference"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets the user preferences",
        "tags": [
          "user-service"
        ]
      },
      "put": {
        "operationId": "updateUserPreferences",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PnPUserPreference"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Updates a user preferences",
        "tags": [
          "user-service"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GrantedUniverseAuthorityDTO": {
        "required": [
          "permission",
          "universe"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/IGrantedAuthorityDTO"
          },
          {
            "type": "object",
            "properties": {
              "permission": {
                "type": "string"
              },
              "universe": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "IGrantedAuthorityDTO": {
        "type": "object"
      },
...
      "PnPUser": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "maxLength": 2147483647,
            "minLength": 3,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "PnPUserCreation": {
        "required": [
          "authorities",
          "displayName",
          "username"
        ],
        "type": "object",
        "properties": {
          "authorities": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/GrantedUniverseAuthorityDTO"
                },
                {
                  "$ref": "#/components/schemas/RoleAuthorityDTO"
                }
              ]
            }
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "PnPUserPreference": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string"
          },
          "lastSelectedUniverse": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "RoleAuthorityDTO": {
        "required": [
          "role"
        ],
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/IGrantedAuthorityDTO"
          },
          {
            "type": "object",
            "properties": {
              "role": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    }
  }
}